Only install yaourt if it is missing

+++ modules/arch/install.sh
@@ -1,20 +1,22 @@
 #!/bin/sh
 
-echo "Installing package-query"
-curl -O https://aur.archlinux.org/packages/pa/package-query/package-query.tar.gz
-tar zxvf package-query.tar.gz
-cd package-query
-makepkg -si --noconfirm
-cd ..
-rm -rf package-query
+if ! hash yaourt 2>/dev/null; then
+    echo "Installing package-query"
+    curl -O https://aur.archlinux.org/packages/pa/package-query/package-query.tar.gz
+    tar zxvf package-query.tar.gz
+    cd package-query
+    makepkg -si --noconfirm
+    cd ..
+    rm -rf package-query
 
-echo "Installing yaourt"
-curl -O https://aur.archlinux.org/packages/ya/yaourt/yaourt.tar.gz
-tar zxvf yaourt.tar.gz
-cd yaourt
-makepkg -si --noconfirm
-cd ..
-rm -rf yaourt
+    echo "Installing yaourt"
+    curl -O https://aur.archlinux.org/packages/ya/yaourt/yaourt.tar.gz
+    tar zxvf yaourt.tar.gz
+    cd yaourt
+    makepkg -si --noconfirm
+    cd ..
+    rm -rf yaourt
+fi
 
 ./write.py "EDITFILES=0" ~/.yaourtrc
 ./write.py "NOCONFIRM=0" ~/.yaourtrc
